Split watch logic to separate file
