+++ src/backends/imap/mod.rs
@@ -1,3 +1,5 @@
 mod imap;
+mod watch;
 
 pub use imap::*;
+pub use watch::*;
